Move Analyze CI job to Buildkite
We only have a single Buildkite agent, but it's a bare metal M1 Mac Mini
provided by MacStadium, so it's considerably faster than the VMs that
run on Azure Pipelines.

So even with the single runner having contention with Danger and Analyze
jobs running serially on it, I think this may help speed up CI overall.